requirements.txt
pillow
aiohttp
python_dotenv
py-cord
logging
aiosqlite